+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Increment the:
 
 ## [Unreleased]
 
+* [TEST] Remove workaround for metrics cardinality limit test
+  [#3663](https://github.com/open-telemetry/opentelemetry-cpp/pull/3663)
+
 ## [1.23 2025-09-25]
 
 * [CodeHealth] Fix clang-tidy warnings part 6