+++ CHANGELOG.md
@@ -18,6 +18,10 @@ Increment the:
 * [SDK] Add bundle version of utf8_range to validate attributes
   [#3512](https://github.com/open-telemetry/opentelemetry-cpp/pull/3512)
 
+* [TEST] Remove workaround for metrics cardinality limit test
+  [#3663](https://github.com/open-telemetry/opentelemetry-cpp/pull/3663)
+
+
 ## [1.23 2025-09-25]
 
 * [CodeHealth] Fix clang-tidy warnings part 6